% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_electionid.R
\name{get_electionid}
\alias{get_electionid}
\title{get_electionid}
\usage{
get_electionid(key = Sys.getenv("google_civic_api"))
}
\arguments{
\item{key}{API key supplied as a string. Defaults to the environmental variable called "google_civic_api." Info on how to get a key can be found here: https://developers.google.com/civic-information}
}
\value{
Returns a list of election informations: id, name, date and ocdDivisionID
}
\description{
This function interacts with the election endpoint of the Google Civic Information API
}
\examples{
\dontrun{
get_electionid(key = Sys.getenv("google_civic_api"))
}
}
