% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rep_address.R
\name{get_rep_address}
\alias{get_rep_address}
\title{get_rep_address}
\usage{
get_rep_address(
  address = NULL,
  includeOffices = NULL,
  levels = NULL,
  roles = NULL,
  key = Sys.getenv("google_civic_api")
)
}
\arguments{
\item{address}{Address supplied as string}

\item{includeOffices}{Whether to return information about offices and officials. False returns only top-level district information. Default is true.}

\item{levels}{List of office levels supplied as string. Only offices that serve at least one of these levels will be returned Acceptable values are: "administrativeArea1", "administrativeArea2", "country", "international", "locality", "regional", "special", "subLocality1", "subLocality2"}

\item{roles}{List of roles supplied as string. Only offices fulfilling one of these roles will be returned. Acceptable values are: "deputyHeadOfGovernment", "executiveCouncil", "governmentOfficer", "headOfGovernment", "headOfState", "highestCourtJudge", "judge", "legislatorLowerBody", "legislatorUpperBody", "schoolBoard", "specialPurposeOfficer"}

\item{key}{API key supplied as a string. Defaults to the environmental variable called "google_civic_api." Info on how to get a key can be found here: https://developers.google.com/civic-information}
}
\value{
Returns a list of offices by divisions and respective officials
}
\description{
This function interacts with the representatives endpoint of the Google Civic Information API
}
\examples{
\dontrun{
get_rep_address(address = "55 e monroe, chicago, il",level="regional", role="schoolBoard", key = Sys.getenv("google_civic_api"))
}

}
