% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_voterinfo.R
\name{get_voterinfo}
\alias{get_voterinfo}
\title{Title}
\usage{
get_voterinfo(
  address = NULL,
  electionid = NULL,
  key = Sys.getenv("google_civic_api")
)
}
\arguments{
\item{electionid}{}
}
\value{

}
\description{
Title
}
