% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocd_divisions.R
\docType{data}
\name{ocd_divisions}
\alias{ocd_divisions}
\title{Open Civic Data Division Identifiers}
\format{
A data frame with 193,493 row and 13 columns, including
\itemize{
\item id
\item name
\item sameAs
\item sameAsNote
\item validThrough
\item census_geoid
\item census_geoid_12
\item census_geoid_14
\item openstates_district
\item placeholder_id
\item sch_dist_stateid
\item state_id
\item validFrom
}
}
\source{
\href{https://github.com/opencivicdata/ocd-division-ids/blob/master/identifiers/country-us.csv}{OpenCivicData}
}
\usage{
ocd_divisions
}
\description{
A data frame of Open Civic Data division identifiers. The IDs are suitable
passing to Google Civic Information API's \code{representativeInfoByDivision}
endpoint. Searching for relevant strings in \code{name} can help locate relevant
IDs.
}
\keyword{datasets}
